[cert-manager] Remove orphan secrets hooks
